Remove files that were generated by an older version of phoenix
When I initially generated the project it was with phoenix 1.16, and
well then I moved it over to 1.17 (because that's the docs that I was
looking at for how to trivially do json), and something wasn't clicking,
so rather than change the docs version that I was looking at, I decided
to use a newer version (and it's great - requires less code), but then
these two were kind of left over from what we had before and all they do
is make initial compilation difficult.

Now of course I had a feeling this would be the case, so after pushing
the repo I, deleted it from my computer, recloned it from github, and
ran my own instructions to verify that it all works on a clean checkout.

Would have been awkward if it didn't 😅